Create Creating_a_java_file.java
Created a new Java file "MyJavaProgram.java" that prints "Hello, World!" to the console. This file was added to the repository and committed with the message "Created a new Java file 'MyJavaProgram.java' that contains the code to print 'Hello, World!' to the console." The file can be compiled and executed using a Java compiler and the Java Virtual Machine (JVM).